% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR_wrapper.R
\name{LRMultiClass}
\alias{LRMultiClass}
\title{Multi-class Logistic Regression}
\usage{
LRMultiClass(X, y, numIter = 50, eta = 0.1, lambda = 1, beta_init = NULL)
}
\arguments{
\item{X}{- is a n x p matrix, where p is the number of covariates.}

\item{y}{- is a vector of length n.}

\item{numIter}{- number of iterations to run.}

\item{eta}{- learning rate for the Damped Newton's Algorithm (> 0).}

\item{lambda}{- shrinkage parameter. For larger lambda more beta coefficients will be shrunk towards zero.}

\item{beta_init}{- starting value for the beta matrix. Must be pxk where k is the number of groups.}
}
\value{
out - the list of beta's & objective values
}
\description{
Multi-class Logistic Regression
}
\examples{
# Give example
data(iris)
# Extract only the features (without labels)
X <- as.matrix(cbind(1, iris[, 1:4]))  # Add intercept
y <- as.integer(factor(iris$Species)) - 1  # Convert species to numeric 0, 1, 2
LRMultiClass(X, y)
plot(LRMultiClass(X, y)$objective, type = 'o')
}
