% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kmeans_wrapper.R
\name{MyKmeans}
\alias{MyKmeans}
\title{Kmeans C++ implementation}
\usage{
MyKmeans(X, K, M = NULL, numIter = 100)
}
\arguments{
\item{X}{- n by p matrix containing n data points to cluster}

\item{K}{- integer specifying number of clusters}

\item{M}{- (optional) K by p matrix of cluster centers}

\item{numIter}{- number of maximal iterations for the algorithm, the default value is 100}
}
\value{
Y - a vector of length n that contains class assignments (0:(K-1)).
}
\description{
Kmeans C++ implementation
}
\examples{
# Give example
set.seed(234) # Set Seed
n <- 100
# Set cluster means:
mean1 <- 1
mean2 <- 5
mean3 <- 10
# Simulate data:
X <- matrix(c(rnorm(n, mean1), rnorm(n, mean2), rnorm(n, mean3)), ncol = 1)
K <- 3
# Run Function:
MyKmeans(X, K)
}
