% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR_wrapper.R
\name{LRMultiClass}
\alias{LRMultiClass}
\title{Multi-class Logistic Regression}
\usage{
LRMultiClass(X, y, numIter = 50, eta = 0.1, lambda = 1, beta_init = NULL)
}
\arguments{
\item{X}{- is a n x p matrix, where p is the number of covariates.}

\item{y}{- is a vector of length n.}

\item{numIter}{- number of iterations to run.}

\item{eta}{- learning rate for the Damped Newton's Algorithm (> 0).}

\item{lambda}{- shrinkage parameter. For larger lambda more beta coefficients will be shrunk towards zero.}

\item{beta_init}{- starting value for the beta matrix. Must be pxk where k is the number of groups.}
}
\value{
A list of the following items:
\itemize{
   \item beta - A matrix (p x K) of the estimated coefficients.
   \item objective - A vector of length num_Iter + 1, that includes the objective function at each iteration.
 }
}
\description{
Multi-class Logistic Regression
}
\examples{
# Give example
set.seed(1234) # set seed
n <- 30 # set n
p <- 3 # set number of covariates
K <- 2 # number of classes
beta <- matrix(c(2, -.11, .1, 1.5, .1, .01), p, K) # set beta
X1 <- cbind(rep(1, n/2), runif(n/2, 2, 3), rnorm(n/2)) # set X1
X2 <- cbind(rep(1, n/2), runif(n/2, -.5, 1), rnorm(n/2, 1)) # set X2
X <- rbind(X1, X2)
exp_XB <- exp(X \%*\% beta)
pk <- exp_XB / rowSums(exp_XB)
y <- apply(pk, 1, which.max) - 1 # Get simulated y:
## Apply Function:
LRMultiClass(X, y)
plot(LRMultiClass(X, y)$objective, type = 'o')
}
