% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kmeans_wrapper.R
\name{MyKmeans}
\alias{MyKmeans}
\title{Kmeans C++ implimentation}
\usage{
MyKmeans(X, K, M = NULL, numIter = 100)
}
\arguments{
\item{X}{- n by p matrix containing n data points to cluster}

\item{K}{- integer specifying number of clusters}

\item{M}{- (optional) K by p matrix of cluster centers}

\item{numIter}{- number of maximal iterations for the algorithm, the default value is 100}
}
\value{
Y - a vector of length n that contains class assignments (0:(K-1)).
}
\description{
Kmeans C++ implimentation
}
\examples{
# Give example
data(iris)
# Extract only the features (without labels)
X <- as.matrix(iris[, 1:4])  # The first 4 columns are features
K <- 3
result <- MyKmeans(X, K)
}
